Fixes #141 for real this time

+++ src/main/java/be/witmoca/BEATs/connection/BackupHandler.java
@@ -107,7 +107,7 @@ class BackupHandler implements ActionListener {
 	}
 
 	private static long countSize(List<File> files) {
-		int i = 0;
+		long i = 0;
 		for (File f : files) {
 			i += f.length();
 		}
